Implemented diamond detection

+++ Test.java
@@ -91,7 +91,7 @@ public class Test {
 		
 		System.out.println();
 		System.out.println("Printing out adjacency matrix. A: ");
-		int[][] A = graph.getAdjacencyMatrix();
+		double[][] A = graph.getAdjacencyMatrix();
 		for(int i=0; i<A.length; i++){
 			for(int j=0; j<A[i].length; j++){
 				System.out.print(A[i][j]+" ");
